+++ src/lib/XmlDocument.ts
@@ -2,6 +2,7 @@ import { XmlElement } from './XmlElement.js';
 import { XmlNode } from './XmlNode.js';
 
 import type { XmlComment } from './XmlComment.js';
+import type { XmlDeclaration } from './XmlDeclaration.js';
 import type { XmlProcessingInstruction } from './XmlProcessingInstruction.js';
 
 /**
@@ -12,9 +13,9 @@ export class XmlDocument extends XmlNode {
   /**
    * Child nodes of this document.
    */
-  readonly children: Array<XmlComment | XmlProcessingInstruction | XmlElement>;
+  readonly children: Array<XmlComment | XmlDeclaration | XmlProcessingInstruction | XmlElement>;
 
-  constructor(children: Array<XmlComment | XmlElement | XmlProcessingInstruction> = []) {
+  constructor(children: Array<XmlComment | XmlDeclaration | XmlElement | XmlProcessingInstruction> = []) {
     super();
     this.children = children;
   }