+++ tests/browser/browserTests.js
@@ -5,6 +5,7 @@ require('../lib/Parser.test.js');
 require('../lib/StringScanner.test.js');
 require('../lib/XmlCdata.test.js');
 require('../lib/XmlComment.test.js');
+require('../lib/XmlDeclaration.test.js');
 require('../lib/XmlDocument.test.js');
 require('../lib/XmlElement.test.js');
 require('../lib/XmlNode.test.js');