+++ tests/index.test.js
@@ -3,11 +3,12 @@
 
 const assert = require('assert');
 
-const { parseXml, XmlCdata, XmlComment, XmlDocument, XmlElement, XmlNode, XmlProcessingInstruction, XmlText } = require('..');
+const { parseXml, XmlCdata, XmlComment, XmlDeclaration, XmlDocument, XmlElement, XmlNode, XmlProcessingInstruction, XmlText } = require('..');
 
 it('exports XML node classes', () => {
   assert.equal(typeof XmlCdata, 'function');
   assert.equal(typeof XmlComment, 'function');
+  assert.equal(typeof XmlDeclaration, 'function');
   assert.equal(typeof XmlDocument, 'function');
   assert.equal(typeof XmlElement, 'function');
   assert.equal(typeof XmlNode, 'function');